Added style to header and hero section
The header's left and right sides are separated with the "space-around" property. It might be best to change this aspect to "space-between" instead and then apply padding to the left and right sides to separate it further from the edge. But for now it will do since I'm designing on a smaller window than what this project is intending.